event_manager/requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile '.\requirements-dev.in'
#
asgiref==3.8.1
    # via
    #   -c C:\Users\Administrator\django_projects\event_project\event_manager\requirements.txt
    #   django
django==5.1.7
    # via
    #   -c C:\Users\Administrator\django_projects\event_project\event_manager\requirements.txt
    #   django-debug-toolbar
    #   django-extensions
django-debug-toolbar==5.1.0
    # via -r .\requirements-dev.in
django-extensions==3.2.3
    # via -r .\requirements-dev.in
factory-boy==3.3.3
    # via -r .\requirements-dev.in
faker==37.1.0
    # via factory-boy
sqlparse==0.5.3
    # via
    #   -c C:\Users\Administrator\django_projects\event_project\event_manager\requirements.txt
    #   django
    #   django-debug-toolbar
tzdata==2025.2
    # via
    #   -c C:\Users\Administrator\django_projects\event_project\event_manager\requirements.txt
    #   django
    #   faker
